fix(core): ignore invalid memory inbox patches
